+++ drillx/cuda/drillx.h
@@ -7,7 +7,7 @@
 #include "equix/src/solver_heap.h"
 #include "hashx/src/context.h"
 
-extern "C" const int BATCH_SIZE;
+extern "C" const uint64_t BATCH_SIZE;
 
 extern "C" void hash(uint8_t *challenge, uint8_t *nonce, uint64_t *out);
